+++ resources/validation.py
@@ -1,6 +1,7 @@
 import subprocess
-from resources import build
-from data import example_data, model_array
+from resources import build, sys_patch_helpers
+from data import example_data, model_array, sys_patch_dict, os_data
+from pathlib import Path
 
 
 def validate(settings):
@@ -58,6 +59,43 @@ def validate(settings):
                 raise Exception(f"Validation failed for predefined model: {settings.computer.real_model}")
             else:
                 print(f"Validation succeeded for predefined model: {settings.computer.real_model}")
+    
+
+    def validate_root_patch_files(major_kernel, minor_kernel):
+        patchset = sys_patch_dict.SystemPatchDictionary(major_kernel, minor_kernel, settings.legacy_accel_support)
+        host_os_float = float(f"{major_kernel}.{minor_kernel}")
+        for patch_subject in patchset:
+            for patch_core in patchset[patch_subject]:
+                patch_os_min_float = float(f'{patchset[patch_subject][patch_core]["OS Support"]["Minimum OS Support"]["OS Major"]}.{patchset[patch_subject][patch_core]["OS Support"]["Minimum OS Support"]["OS Minor"]}')
+                patch_os_max_float = float(f'{patchset[patch_subject][patch_core]["OS Support"]["Maximum OS Support"]["OS Major"]}.{patchset[patch_subject][patch_core]["OS Support"]["Maximum OS Support"]["OS Minor"]}')
+                if (host_os_float < patch_os_min_float or host_os_float > patch_os_max_float):
+                    continue
+                for install_type in ["Install", "Install Non-Root"]:
+                    if install_type in patchset[patch_subject][patch_core]:
+                        for install_directory in patchset[patch_subject][patch_core][install_type]:
+                            for install_file in patchset[patch_subject][patch_core][install_type][install_directory]:
+                                source_file = str(settings.payload_local_binaries_root_path) + "/" + patchset[patch_subject][patch_core][install_type][install_directory][install_file] + install_directory + "/" + install_file
+                                if not Path(source_file).exists():
+                                    raise Exception(f"Failed to find {source_file}")
+        
+        print(f"Validating Root Patch Dictionary integrity for Darwin {major_kernel}.{minor_kernel}")
+        if not sys_patch_helpers.sys_patch_helpers(settings).generate_patchset_plist(patchset, "OpenCore-Legacy-Patcher"):
+            raise Exception("Failed to generate patchset plist")
+
+    
+    def validate_sys_patch():
+        if Path(settings.payload_local_binaries_root_path_zip).exists():
+            print("Validating Root Patch File integrity")
+            if not Path(settings.payload_local_binaries_root_path).exists():
+                subprocess.run(["ditto", "-V", "-x", "-k", "--sequesterRsrc", "--rsrc", settings.payload_local_binaries_root_path_zip, settings.payload_path])
+
+            for supported_os in [os_data.os_data.big_sur, os_data.os_data.monterey]:
+                validate_root_patch_files(supported_os, 6)
+        print("Validating SNB Board ID patcher")
+            settings.computer.reported_board_id = "Mac-7BA5B2DFE22DDD8C"
+            sys_patch_helpers.sys_patch_helpers(settings).snb_board_id_patch(settings.payload_local_binaries_root_path)
+        else:
+            print("Skipping Root Patch File integrity validation")
 
     # First run is with default settings
     build_prebuilt()
@@ -81,3 +119,4 @@ def validate(settings):
     settings.serial_settings = "Minimal"
     build_prebuilt()
     build_dumps()
+    validate_sys_patch()